middleware/l3/python/requirements.txt
zeroc-ice==3.7.9